Client Updates
Modified the client and added the playInterface.
Modified functionality to represent the sub systems defined in the specification doc.
The client is set take in messages as defined in the message class and has been tested using stubs.
At this point the client just sends text back to the server but this can be updated.