fixed issue with cycle object
issue with cycler determined to have been caused by player_dict being modified after cycler was instantiated.  Because the instantiation of cycler contains a reference to the player_dict object, this meant that updating the player dict would break cycler.  This was solved by adding an opt-in prompt after the initial connection by the clients to the server but before the response to the initial query message was sent by the client, and moving cycler's instantiation to occur only after the opt-ins were all accepted.  This essentially ensured that cycler was blocked from instantiating until player_list was locked in.

Currently, however, any player opting in to ready would remove the block and open cycler up to be instantiated, so no player should enter ready until all players playing have connected.

One issue remains, the first player to opt-in is given two moves at the start, it's unclear currently why this occurs, because I'm quite sure that sending Move the first time should rotate cycler.  I'm really not sure why that happens.

However, asides from that, the server is able to cycle prompts between players as it should for the skeletal increment.